+++ bomoseries-Disney-api/bomoseries-Disney-api/Watchable.cs
@@ -6,6 +6,6 @@
         public string Title { get; set; } = string.Empty;
         public string Type { get; set; } = string.Empty;
         public string Date { get; set; } = string.Empty;
-        public float Rating { get; set; }
+        public double Rating { get; set; }
     }
 }
